Update siddhi-4.0.md
Changing the line number 893 from 
Following is a Siddhi App the controls the temperature regulators if they are not already `on` for all the rooms with a room temperature greater than 30 degrees.  
to 
Following is a Siddhi App that controls the temperature regulators if they are not already `on` for all the rooms with a room temperature greater than 30 degrees.